requirements.unstable.txt
beautifulsoup
bleach
boto
django-celery-with-redis
django<1.6
django-analytical>=0.12.1
django-annoying
django-allauth
django_compressor
django-heroku-memcacheify
django-heroku-postgresify
django-longerusernameandemail
django-extensions
django-mailgun
django-nose
django-selector
django-sslify
django-staticfiles
django-storages
django-ses
dropbox
fabric
facepy
gunicorn
mistune
pillow
psycopg2
requests
requests-oauth2
titlecase
tweepy
sorl-thumbnail
south
splinter
versiontools

# Dev only
django-debug-toolbar
ipython